move optional.py
